[feature] adding endpoint selection feature

+++ lib/grape/resources.rb
@@ -5,17 +5,17 @@ module Grape
   class API
     include Grape::Resources
     class << self
-      def resources_for( clazz )
+      def resources_for( clazz, methods=[:list, :get, :post, :put, :delete])
         singular_name = clazz.name.underscore
         plural_name   = clazz.name.pluralize.underscore
 
         raise Error("To use grape_resources on a given class it should inherit from ActiveRecord::Base.( at least for now buddy ;) )") unless clazz < ActiveRecord::Base
 
-        Grape::Resources.list_endpoint_for( clazz, self )
-        Grape::Resources.get_endpoint_for( clazz, self )
-        Grape::Resources.create_endpoint_for( clazz, self )        
-        Grape::Resources.update_endpoint_for( clazz, self )                
-        Grape::Resources.delete_endpoint_for(clazz, self)
+        Grape::Resources.list_endpoint_for( clazz, self ) if methods.include?(:list)
+        Grape::Resources.get_endpoint_for( clazz, self ) if methods.include?(:get)
+        Grape::Resources.create_endpoint_for( clazz, self ) if methods.include?(:post)        
+        Grape::Resources.update_endpoint_for( clazz, self ) if methods.include?(:put)               
+        Grape::Resources.delete_endpoint_for(clazz, self) if methods.include?(:delete)
       end
     end
   end
